* Feature: (Premium) Advanced Comment Spam Filter. Checks comment source IP, author URL and hosts and IP's in body against additional spam lists. * Feature: (Premium) Check if your site is being Spamvertised i.e. your domain is being included in spam emails. Usually indicates you've been hacked. * Feature: (Premium) Check if your website IP is generating spam. Checks against spam lists if your IP is a known source of spam. * Improvement: Cache clearing errors are nown shown with clear explanations. * Improvement: Added lightweight stats logging internally in preparation for displaying them on the admin UI in the next release. * Fix: If a non-existent user tries to sign in it is not logged in the live logins tab. Fixed. * Fix: Removed warning "Trying to get property of non-object" that would occur under certain conditions. * Fix: Removed call to is_404() which was not having any effect and would issue a warning if debug mode is enabled. * Fix: Check if CURL is installed as part of connectivity test.
=


git-svn-id: https://plugins.svn.wordpress.org/wordfence/trunk@922566 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/conntest.php
@@ -40,6 +40,10 @@ function doWPostTest($protocol){
 	}
 }
 function doCurlTest($protocol){
+	if(! function_exists('curl_init')){
+		echo "<br /><b style='color: #F00;'>CURL is not installed</b>. Asking your hosting provider to install and enable CURL may improve any connection problems.</b><br />\n";
+		return;
+	}
 	echo "<br /><b>STARTING CURL $protocol CONNECTION TEST....</b><br />\n";
 	global $curlContent;
 	$curlContent = "";
